Add: Data Validation (Stock, Comment)

+++ Controllers/CommentController.cs
@@ -12,48 +12,72 @@ namespace api.Controllers
 {
     [Route("api/comment")]
     [ApiController]
-    public class CommentController:ControllerBase
+    public class CommentController : ControllerBase
     {
         private readonly ICommentRepository _commentRepository;
         private readonly IStockRepository _stockRepository;
-        public CommentController(ICommentRepository commentRepository, IStockRepository stockRepository){
+        public CommentController(ICommentRepository commentRepository, IStockRepository stockRepository)
+        {
             _commentRepository = commentRepository;
             _stockRepository = stockRepository;
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll(){
+        public async Task<IActionResult> GetAll()
+        {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepository.GetAllAsync();
             var commentDto = comments.Select(s => s.ToCommentDto());
             return Ok(commentDto);
         }
 
         [HttpGet]
-        [Route("{id}")]
-        public async Task<IActionResult> GetById([FromRoute] int id){
+        [Route("{id:int}")]
+        public async Task<IActionResult> GetById([FromRoute] int id)
+        {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comment = await _commentRepository.GetByIdAsync(id);
-            if(comment == null){
+            if (comment == null)
+            {
                 return NotFound();
             }
 
             return Ok(comment.ToCommentDto());
         }
-        
-        [HttpPost("{stockId}")]
-        public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto){
-            if(!await _stockRepository.StockExists(stockId)){
+
+        [HttpPost("{stockId}:int")]
+        public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
+        {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            if (!await _stockRepository.StockExists(stockId))
+            {
                 return BadRequest("Stock does not exist");
             }
             var commentModel = commentDto.ToCommentFromCreate(stockId);
             await _commentRepository.CreateAsync(commentModel);
-            return CreatedAtAction(nameof(GetById), new{id = commentModel.Id}, commentModel.ToCommentDto());
+            return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
-        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto){
+        [Route("{id:int}")]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
+        {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comment = await _commentRepository.UpdateAsync(id, updateDto.ToCommentFromUpdate());
-            if(comment == null){
+            if (comment == null)
+            {
                 return NotFound("Comment not found");
             }
 
@@ -61,10 +85,16 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
-        public async Task<IActionResult> Delete([FromRoute] int id){
+        [Route("{id:int}")]
+        public async Task<IActionResult> Delete([FromRoute] int id)
+        {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var commentModel = await _commentRepository.DeleteAsync(id);
-            if(commentModel == null){
+            if (commentModel == null)
+            {
                 return NotFound("Comment does not exist");
             }
             return Ok(commentModel.ToCommentDto());
